Fix several bugs in the scooer reservation system

+++ backend/src/services/reservation-manager.ts
@@ -6,13 +6,19 @@ import { CronJob } from 'cron';
 
 abstract class ReservationManager {
     /* check if a scooter is reserved */
-    public static async getReservationFromScooter(scooterId: number): Promise<Model> {
-        return await Reservation.findOne({ where: { scooter_id: scooterId } });
+    public static async getReservationFromScooter(scooterId: number, transaction?: Transaction): Promise<Model> {
+      if (transaction) {
+        return await Reservation.findOne({ where: { scooter_id: scooterId }, transaction: transaction });
+      }
+      return await Reservation.findOne({ where: { scooter_id: scooterId } });
     }
 
     /* check if a user has a reservation */
-    public static async getReservationFromUser(userId: number): Promise<Model> {
-        return await Reservation.findOne({ where: {user_id: userId } });
+    public static async getReservationFromUser(userId: number, transaction?: Transaction): Promise<Model> {
+      if (transaction) {
+        return await Reservation.findOne({ where: {user_id: userId }, transaction: transaction });
+      }
+      return await Reservation.findOne({ where: {user_id: userId } });
     }
 
     // set a scooter as reserved for a user, if possible (scooter is free and user has no other reservations)
@@ -22,19 +28,23 @@ abstract class ReservationManager {
         let expiration: Date;
         const transactionExtern: boolean = transaction !== undefined;
 
-        // can't reserve if the scooter isn't real
-        const scooter = await Scooter.findByPk(scooterId);
-        if(!scooter) throw new Error('SCOOTER_DOES_NOT_EXIST');
-        // can't reserve if scooter is reserved or rented
-        if(scooter.getDataValue('active_rental_id') !== null || scooter.getDataValue('reservation_id') !== null || await this.getReservationFromScooter(scooterId) !== null) {
-            throw new Error('SCOOTER_UNAVAILABLE');
-        }
-        // can't reserve if user already has reservation
-        if(await this.getReservationFromUser(userId)) {
-            throw new Error('USER_HAS_RESERVATION');
-        }
         if(!transactionExtern) transaction = await database.getSequelize().transaction();
+
         try {
+            // can't reserve if the scooter isn't real
+            const scooter = await Scooter.findByPk(scooterId, { transaction: transaction });
+            if(!scooter) throw new Error('SCOOTER_DOES_NOT_EXIST');
+
+            // can't reserve if scooter is reserved or rented
+            if(scooter.getDataValue('active_rental_id') !== null || scooter.getDataValue('reservation_id') !== null || await this.getReservationFromScooter(scooterId) !== null) {
+                throw new Error('SCOOTER_UNAVAILABLE');
+            }
+
+            // can't reserve if user already has reservation
+            if(await this.getReservationFromUser(userId)) {
+                throw new Error('USER_HAS_RESERVATION');
+            }
+
             // all good?
             // create the new reservation
             expiration = new Date(Date.now() + RESERVATION_LIFETIME);
@@ -66,12 +76,18 @@ abstract class ReservationManager {
 
     // end a reservation, freeing the scooter and user
     // if caller provides a transaction, use that and don't commit/rollback. otherwise, checkout and manage a new transaction
-    public static async endReservation(reservation: Model, transaction?: Transaction): Promise<void> {
+    public static async endReservation(reservation: Model, transaction?: Transaction, scooter?: Model): Promise<void> {
         const transactionExtern: boolean = transaction !== undefined;
         if(!transactionExtern) transaction = await database.getSequelize().transaction();
         try {
-            const scooter = await Scooter.findByPk(reservation.getDataValue('scooter_id'));
-            await reservation.destroy({transaction: transaction});
+            /* Fetch the scooter if it wasn't provided */
+            if (!scooter) {
+              scooter = await Scooter.findByPk(reservation.getDataValue('scooter_id'), { transaction: transaction });
+            }
+            if(!scooter) throw new Error('SCOOTER_NOT_FOUND');
+
+            // const scooter = await Scooter.findByPk(reservation.getDataValue('scooter_id'));
+            await reservation.destroy({ transaction: transaction });
             scooter.setDataValue('reservation_id', null);
             await scooter.save({transaction: transaction});
             if(!transactionExtern) await transaction.commit();
